Characters now have different default emotions.
Before, a character's default emotion HAD to be called "n". While that gelled with Enthalpy's suffix-naming system, it doesn't give people a lot of freedom.
If I wanted to make a posthumous character who only had one of those dead-people signs, I'd have to give that sprite the suffix "n", and that's just counterintuitive. I like this better.